+++ src/plugins/playlists.js
@@ -9,8 +9,7 @@ import {
 	deleteSongFromPlaylistByIdHandler,
 	getPlaylistActivitiesHandler,
 } from "../handlers/playlists.js";
-import { InvariantError } from "../utils/error/invariant-error.js";
-import { NotFoundError } from "../utils/error/notfound-error.js";
+import { InvariantError, NotFoundError } from "../utils/error.js";
 
 const playlistBodySchema = Joi.object({
 	name: Joi.string().trim().required(),
