+++ src/server.js
@@ -11,13 +11,15 @@ import { authPlugin } from "./plugins/authentications.js";
 import { usersPlugin } from "./plugins/users.js";
 import { playlistsPlugin } from "./plugins/playlists.js";
 import { collabPlugin } from "./plugins/collaborations.js";
+
 import { albumsServicePlugin } from "./plugins/albums-service.js";
 import { songsServicePlugin } from "./plugins/songs-service.js";
 import { usersServicePlugin } from "./plugins/users-service.js";
 import { authServicePlugin } from "./plugins/authentications-service.js";
 import { playlistsServicePlugin } from "./plugins/playlists-service.js";
 import { collabServicePlugin } from "./plugins/collaborations-service.js";
-import { ClientError } from "./utils/error/client-error.js";
+
+import { ClientError } from "./utils/error.js";
 
 dotenv.config();
